+++ packages/blocks/src/BrandLogos/types.ts
@@ -1,9 +1,8 @@
-import { TextPairingVariant, TextVariant, Theme } from '@cmpsr/components';
+import { TextPairingProps, TextProps, Theme } from '@cmpsr/components';
 
 export interface BrandLogosProps {
   backgroundColor?: ThisType<Theme['colors']>;
-  title?: string;
-  description?: string;
   logos: Array<string>;
-  textVariant?: TextVariant | TextPairingVariant;
+  logosMaxHeight?: string;
+  textProps: TextPairingProps | TextProps;
 }