+++ packages/nextjs-contentful-renderer/src/getPageContent/getPageContent.ts
@@ -1,3 +1,4 @@
+import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
 import { GetServerSidePropsContext } from 'next';
 import { configNavbar } from '../configNavbar';
 import { configFooter } from '../configFooter';
@@ -20,7 +21,7 @@ export const getPageContent = async (
 
   const existingPageId = getVisitedPageIdFromCookies(context, slug);
   if (existingPageId) {
-    const page = await getPageById(apolloClient, existingPageId, preview);
+    const page = await getPageByIdAndConfigNavbar(apolloClient, existingPageId, preview);
     if (page) {
       return page;
     }
@@ -30,13 +31,20 @@ export const getPageContent = async (
   if (!route) return undefined;
 
   const pageId = getPageId(route, context.query.utm_campaign);
-  const pageContent = await getPageById(apolloClient, pageId, preview);
+  const page = await getPageByIdAndConfigNavbar(apolloClient, pageId, preview);
   setCookie(context, slug, pageId);
 
-  if (pageContent) {
-    pageContent.content = configNavbar(pageContent);
-    pageContent.content = configFooter(pageContent);
-  }
+  return page;
+};
 
-  return pageContent;
+const getPageByIdAndConfigNavbar = async (
+  apolloClient: ApolloClient<NormalizedCacheObject>,
+  pageId: string,
+  preview: boolean
+) => {
+  const page = await getPageById(apolloClient, pageId, preview);
+  if (page) {
+    page.content = configNavbar(page);
+  }
+  return page;
 };