+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @cmpsr/blocks
 
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [c06accb]
+  - @cmpsr/components@3.1.0
+
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies [f292607]
+  - @cmpsr/components@3.0.0
+
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [adba0b1]
+  - @cmpsr/components@2.4.0
+
+## 3.3.0
+
+### Minor Changes
+
+- d09879f: feat: unify image properties into imageProps
+
+## 3.2.0
+
+### Minor Changes
+
+- c44391d: feat: create new Hero block, add maxHeight for brand logos, allow logoHref for navigation, allow linkProps for testimonial
+
 ## 3.1.0
 
 ### Minor Changes